requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
Automat==20.2.0
beautifulsoup4==4.11.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==37.0.4
cssselect==1.1.0
Django==3.2.15
djangorestframework==3.13.1
filelock==3.8.0
h11==0.13.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.7.0
itemloaders==1.0.5
jmespath==1.0.1
lxml==4.9.1
outcome==1.2.0
parsel==1.6.0
pathlib==1.0.1
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
PyMySQL==1.0.2
pyOpenSSL==22.0.0
PySocks==1.7.1
pytz==2022.2.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.6.2
selenium==4.4.3
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tldextract==3.3.1
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typing_extensions==4.3.0
urllib3==1.26.12
w3lib==2.0.1
wsproto==1.2.0
zope.interface==5.4.0
